Rename file wquadtree -> qbvh
